# docs/openapi/custom-endpoints.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "description": "Generated inventory of every repository-owned Backend and Storefront route. Detailed schemas for security-sensitive public boundaries remain in api-problems.yaml.",
    "title": "Remorseless Records custom endpoint inventory",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/admin/catalog/media/orphans": {
      "get": {
        "operationId": "backend_get_admin_catalog_media_orphans",
        "summary": "GET /admin/catalog/media/orphans",
        "tags": [
          "backend custom routes"
        ],
        "responses": {
          "2XX": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "A successful custom-route response.",
            "headers": {
              "X-Request-Id": {
                "$ref": "./api-problems.yaml#/components/headers/RequestId"
              },
              "traceparent": {
                "$ref": "./api-problems.yaml#/components/headers/Traceparent"
              }
            }
          },
          "4XX": {
            "$ref": "./api-problems.yaml#/components/responses/NativeMedusaErrorResponse"
          },
          "5XX": {
            "$ref": "./api-problems.yaml#/components/responses/NativeMedusaErrorResponse"
          }
        },
        "x-error-contract": {
          "envelope": "native-medusa-error",
          "providerFailure": "not-applicable",
          "unexpectedStatus": 500
        },
        "x-route-files": [
          "backend/src/api/admin/catalog/media/orphans/route.ts"
        ],
        "x-services": [
          "backend"
        ]
      }
    },
...
    "/admin/refund-operations": {
      "get": {
        "operationId": "backend_get_admin_refund_operations",
        "summary": "GET /admin/refund-operations",
        "tags": [
          "backend custom routes"
        ],
        "responses": {
          "2XX": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "A successful custom-route response.",
            "headers": {
              "X-Request-Id": {
                "$ref": "./api-problems.yaml#/components/headers/RequestId"
              },
              "traceparent": {
                "$ref": "./api-problems.yaml#/components/headers/Traceparent"
              }
            }
          },
          "4XX": {
            "$ref": "./api-problems.yaml#/components/responses/NativeMedusaErrorResponse"
          },
          "5XX": {
            "$ref": "./api-problems.yaml#/components/responses/NativeMedusaErrorResponse"
          }
        },
        "x-error-contract": {
          "envelope": "native-medusa-error",
          "providerFailure": "not-applicable",
          "unexpectedStatus": 500
        },
        "x-route-files": [
          "backend/src/api/admin/refund-operations/route.ts"
        ],
        "x-services": [
          "backend"
        ]
      }
    },
    "/admin/tax-control": {
      "get": {
        "operationId": "backend_get_admin_tax_control",
        "summary": "GET /admin/tax-control",
        "tags": [
          "backend custom routes"
        ],
        "responses": {
          "2XX": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "A successful custom-route response.",
            "headers": {
              "X-Request-Id": {
                "$ref": "./api-problems.yaml#/components/headers/RequestId"
              },
              "traceparent": {
                "$ref": "./api-problems.yaml#/components/headers/Traceparent"
              }
            }
          },
          "4XX": {
            "$ref": "./api-problems.yaml#/components/responses/NativeMedusaErrorResponse"
          },
          "5XX": {
            "$ref": "./api-problems.yaml#/components/responses/NativeMedusaErrorResponse"
          }
        },
        "x-error-contract": {
          "envelope": "native-medusa-error",
          "providerFailure": "not-applicable",
          "unexpectedStatus": 500
        },
        "x-route-files": [
          "backend/src/api/admin/tax-control/route.ts"
        ],
        "x-services": [
          "backend"
        ]
      }
    },
    "/admin/tax-records": {
      "get": {
        "operationId": "backend_get_admin_tax_records",
        "summary": "GET /admin/tax-records",
        "tags": [
          "backend custom routes"
        ],
        "responses": {
          "2XX": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "A successful custom-route response.",
            "headers": {
              "X-Request-Id": {
                "$ref": "./api-problems.yaml#/components/headers/RequestId"
              },
              "traceparent": {
                "$ref": "./api-problems.yaml#/components/headers/Traceparent"
              }
            }
          },
          "4XX": {
            "$ref": "./api-problems.yaml#/components/responses/ApiProblemResponse"
          },
          "5XX": {
            "$ref": "./api-problems.yaml#/components/responses/ApiProblemResponse"
          }
        },
        "x-error-contract": {
          "envelope": "api-problem",
          "providerFailure": "not-applicable",
          "unexpectedStatus": 500
        },
        "x-route-files": [
          "backend/src/api/admin/tax-records/route.ts"
        ],
        "x-services": [
          "backend"
        ]
      }
    },
    "/admin/tax-records/export": {
      "get": {
        "operationId": "backend_get_admin_tax_records_export",
        "summary": "GET /admin/tax-records/export",
        "tags": [
          "backend custom routes"
        ],
        "responses": {
          "2XX": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "A successful custom-route response.",
            "headers": {
              "X-Request-Id": {
                "$ref": "./api-problems.yaml#/components/headers/RequestId"
              },
              "traceparent": {
                "$ref": "./api-problems.yaml#/components/headers/Traceparent"
              }
            }
          },
          "4XX": {
            "$ref": "./api-problems.yaml#/components/responses/ApiProblemResponse"
          },
          "5XX": {
            "$ref": "./api-problems.yaml#/components/responses/ApiProblemResponse"
          }
        },
        "x-error-contract": {
          "envelope": "api-problem",
          "providerFailure": "not-applicable",
          "unexpectedStatus": 500
        },
        "x-route-files": [
          "backend/src/api/admin/tax-records/export/route.ts"
        ],
        "x-services": [
          "backend"
        ]
      }
    },
...
    "/admin/catalog/media/uploads": {
      "post": {
        "operationId": "backend_post_admin_catalog_media_uploads",
        "summary": "POST /admin/catalog/media/uploads",
        "tags": [
          "backend custom routes"
        ],
        "responses": {
          "2XX": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "A successful custom-route response.",
            "headers": {
              "X-Request-Id": {
                "$ref": "./api-problems.yaml#/components/headers/RequestId"
              },
              "traceparent": {
                "$ref": "./api-problems.yaml#/components/headers/Traceparent"
              }
            }
          },
          "4XX": {
            "$ref": "./api-problems.yaml#/components/responses/NativeMedusaErrorResponse"
          },
          "5XX": {
            "$ref": "./api-problems.yaml#/components/responses/NativeMedusaErrorResponse"
          }
        },
        "x-error-contract": {
          "envelope": "native-medusa-error",
          "providerFailure": "not-applicable",
          "unexpectedStatus": 500
        },
        "x-route-files": [
          "backend/src/api/admin/catalog/media/uploads/route.ts"
        ],
        "x-services": [
          "backend"
        ]
      }
    },
...
    "/admin/managed-uploads": {
      "post": {
        "operationId": "backend_post_admin_managed_uploads",
        "summary": "POST /admin/managed-uploads",
        "tags": [
          "backend custom routes"
        ],
        "responses": {
          "2XX": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "A successful custom-route response.",
            "headers": {
              "X-Request-Id": {
                "$ref": "./api-problems.yaml#/components/headers/RequestId"
              },
              "traceparent": {
                "$ref": "./api-problems.yaml#/components/headers/Traceparent"
              }
            }
          },
          "4XX": {
            "$ref": "./api-problems.yaml#/components/responses/NativeMedusaErrorResponse"
          },
          "5XX": {
            "$ref": "./api-problems.yaml#/components/responses/NativeMedusaErrorResponse"
          }
        },
        "x-error-contract": {
          "envelope": "native-medusa-error",
          "providerFailure": "not-applicable",
          "unexpectedStatus": 500
        },
        "x-route-files": [
          "backend/src/api/admin/managed-uploads/route.ts"
        ],
        "x-services": [
          "backend"
        ]
      }
    },
...
    "/admin/tax-control/switch": {
      "post": {
        "operationId": "backend_post_admin_tax_control_switch",
        "summary": "POST /admin/tax-control/switch",
        "tags": [
          "backend custom routes"
        ],
        "responses": {
          "2XX": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "A successful custom-route response.",
            "headers": {
              "X-Request-Id": {
                "$ref": "./api-problems.yaml#/components/headers/RequestId"
              },
              "traceparent": {
                "$ref": "./api-problems.yaml#/components/headers/Traceparent"
              }
            }
          },
          "4XX": {
            "$ref": "./api-problems.yaml#/components/responses/NativeMedusaErrorResponse"
          },
          "5XX": {
            "$ref": "./api-problems.yaml#/components/responses/NativeMedusaErrorResponse"
          }
        },
        "x-error-contract": {
          "envelope": "native-medusa-error",
          "providerFailure": "not-applicable",
          "unexpectedStatus": 500
        },
        "x-route-files": [
          "backend/src/api/admin/tax-control/switch/route.ts"
        ],
        "x-services": [
          "backend"
        ]
      }
    },
    "/admin/tax-control/taxrate-io/refresh": {
      "post": {
        "operationId": "backend_post_admin_tax_control_taxrate_io_refresh",
        "summary": "POST /admin/tax-control/taxrate-io/refresh",
        "tags": [
          "backend custom routes"
        ],
        "responses": {
          "2XX": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "A successful custom-route response.",
            "headers": {
              "X-Request-Id": {
                "$ref": "./api-problems.yaml#/components/headers/RequestId"
              },
              "traceparent": {
                "$ref": "./api-problems.yaml#/components/headers/Traceparent"
              }
            }
          },
          "4XX": {
            "$ref": "./api-problems.yaml#/components/responses/NativeMedusaErrorResponse"
          },
          "5XX": {
            "$ref": "./api-problems.yaml#/components/responses/NativeMedusaErrorResponse"
          }
        },
        "x-error-contract": {
          "envelope": "native-medusa-error",
          "providerFailure": "not-applicable",
          "unexpectedStatus": 500
        },
        "x-route-files": [
          "backend/src/api/admin/tax-control/taxrate-io/refresh/route.ts"
        ],
        "x-services": [
          "backend"
        ]
      }
    },
...
    "/webhooks/stripe/lifecycle": {
      "post": {
        "operationId": "backend_post_webhooks_stripe_lifecycle",
        "summary": "POST /webhooks/stripe/lifecycle",
        "tags": [
          "backend custom routes"
        ],
        "responses": {
          "2XX": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "A successful custom-route response.",
            "headers": {
              "X-Request-Id": {
                "$ref": "./api-problems.yaml#/components/headers/RequestId"
              },
              "traceparent": {
                "$ref": "./api-problems.yaml#/components/headers/Traceparent"
              }
            }
          },
          "4XX": {
            "$ref": "./api-problems.yaml#/components/responses/ApiProblemResponse"
          },
          "5XX": {
            "$ref": "./api-problems.yaml#/components/responses/ApiProblemResponse"
          }
        },
        "x-error-contract": {
          "envelope": "api-problem",
          "providerFailure": "not-applicable",
          "unexpectedStatus": 500
        },
        "x-route-files": [
          "backend/src/api/webhooks/stripe/lifecycle/route.ts"
        ],
        "x-services": [
          "backend"
        ]
      }
    },
...
  },
...
}